+++ src/main/java/io/github/astrapi69/runtime/compiler/SeparatorConstants.java
@@ -1,0 +1,60 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2015 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package io.github.astrapi69.runtime.compiler;
+
+/**
+ * The Class SeparatorConstants.
+ */
+public class SeparatorConstants
+{
+
+	/** The Constant WHITE_SPACE. */
+	public static final String WHITE_SPACE = " ";
+
+	/** The Constant DOT. */
+	public static final String DOT = ".";
+
+	/** The Constant DOUBLE_DOT. */
+	public static final String DOUBLE_DOT = ":";
+
+	/** The Constant COLON. */
+	public static final String COLON = ":";
+
+	/** The Constant COLON_WHITE_SPACE. */
+	public static final String COLON_WHITE_SPACE = COLON + WHITE_SPACE;
+
+	/** The Constant SEMI_COLON. */
+	public static final String SEMI_COLON = ";";
+
+	/** The Constant SEMI_COLON_WHITE_SPACE. */
+	public static final String SEMI_COLON_WHITE_SPACE = SEMI_COLON + WHITE_SPACE;
+
+	/** The Constant SLASH. */
+	public static final String SLASH = "/";
+
+	/** The Constant BACK_SLASH. */
+	public static final String BACK_SLASH = "\\";
+
+}