+++ src/test/java/io/github/astrapi69/runtime/compiler/JavaSourceCompilerTest.java
@@ -1,0 +1,55 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2015 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package io.github.astrapi69.runtime.compiler;
+
+import io.github.astrapi69.runtime.compiler.JavaSourceCompiler;
+import org.testng.annotations.Test;
+
+/**
+ * The unit test class for the class {@link JavaSourceCompiler}.
+ *
+ */
+public class JavaSourceCompilerTest
+{
+
+	/**
+	 * Test for method {@link JavaSourceCompiler#compile(String, String, String)}
+	 *
+	 * @throws InstantiationException
+	 *             the instantiation exception
+	 * @throws IllegalAccessException
+	 *             the illegal access exception
+	 */
+	@Test(enabled = true)
+	public void testCompile() throws InstantiationException, IllegalAccessException
+	{
+		final JavaSourceCompiler<Runnable> runtimeCompiler = new JavaSourceCompiler<>();
+		final String source = "public final class FooRunnable implements Runnable { public void run() { System.out.println(\"Foo bar bla\"); } } ";
+		final Class<Runnable> clazz = runtimeCompiler.compile(null, "FooRunnable", source);
+		final Runnable r = clazz.newInstance();
+		r.run();
+	}
+
+}