+++ CHANGELOG.md
@@ -4,9 +4,16 @@
 Version 1.3-SNAPSHOT
 -------------
 
+ADDED:
+
+- gradle as new build system
+
 CHANGED:
 
-- update of parent version to 4.5
+- changed to new package io.github.astrapi69
+- update of dependency lombok to new version 1.18.18
+- update of dependency commons-lang3 version to 3.12.0
+- update of test dependency testng to new version 7.4.0
 
 Version 1.2
 -------------